requirements.txt
streamlit>=1.22
requests>=2.28
python-dotenv>=1.0
Pillow>=9.0